requirements.txt
alembic==0.8.4
blinker==1.4
extras==0.0.3
Flask==0.10.1
Flask-API==0.6.4
Flask-Fixtures==0.3.3
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
iniherit==0.3.4
itsdangerous==0.24
Jinja2==2.8
linecache2==1.0.0
Mako==1.0.3
MarkupSafe==0.23
MySQL-python==1.2.5
passlib==1.6.5
pbr==1.8.1
python-dateutil==2.4.2
python-editor==0.5
python-mimeparse==0.1.4
python-slugify==1.1.4
six==1.10.0
SQLAlchemy==1.0.11
testtools==1.8.1
traceback2==1.4.0
Unidecode==0.4.18
unittest2==1.1.0
Werkzeug==0.11.3
wheel==0.24.0
WTForms==2.1
